+++ src/test/java/net/tab/NetTab.java
@@ -1,0 +1,8 @@
+package net.tab;
+
+/**
+ * This is a placeholder class for testing the version
+ * utility
+ */
+public class NetTab {
+}